Updates module configuration and dependencies
Updates the module configuration to include only the root module, removing references to sub-modules.

Adds a provided scope for log4j-core dependency in SIRE-Commons to avoid conflicts and ensures that the dependency is available at compile time as well.

Includes log4j-api as a compile-time dependency in SIRE-SRI-Reports and SIRE-Services.